upgraded mumbai, improved related scripts
